создает TripController и переносит туда рендер событий
