+++ src/main.js
@@ -3,80 +3,14 @@ import MenuComponent from "./components/menu";
 import FilterComponent from "./components/filter";
 import SortComponent from "./components/sort";
 import BoardComponent from "./components/board";
-import DayComponent from "./components/event-day";
-import EventsListComponent from "./components/events";
-import EventComponent from "./components/event";
-import EventEditComponent from "./components/event-edit";
 import NoEventComponent from "./components/no-event";
+import TripController from "./controllers/tripControllers";
 
-import {generateDay, generateEvent, generateEvents} from "./Mocks/event-mock";
+import {generateDay, generateEvents} from "./Mocks/event-mock";
 
 import {EvenOption, Place} from "./components/consts";
 
-import {getRandomIntegerNumber} from "./utils/common";
-import {render, replace} from "./utils/render";
-
-const replaceEventToEdit = (eventComponent, eventEditComponent) => {
-  openEvent = true;
-  replace(eventEditComponent, eventComponent);
-};
-
-const replaceEditToTask = (eventComponent, eventEditComponent) => {
-  replace(eventComponent, eventEditComponent);
-  openEvent = null;
-};
-
-const renderEvent = (container, event) => {
-  const onEscKeyDown = (evt) => {
-    const isEscKey = evt.key === `Escape` || evt.key === `Esc`;
-
-    if (isEscKey) {
-      replaceEditToTask(eventComponent, eventEditComponent);
-      document.removeEventListener(`keydown`, onEscKeyDown);
-      openEvent = null;
-    }
-  };
-
-  const onReplaceToEdit = () => {
-    if (!openEvent) {
-      replaceEventToEdit(eventComponent, eventEditComponent);
-      document.addEventListener(`keydown`, onEscKeyDown);
-    }
-  };
-
-  const onReplaceToEvent = (evt) => {
-    evt.preventDefault();
-    replaceEditToTask(eventComponent, eventEditComponent);
-    document.removeEventListener(`keydown`, onEscKeyDown);
-  };
-
-  const eventComponent = new EventComponent(event);
-
-  eventComponent.setEditBtnClickHandler(onReplaceToEdit);
-
-  const eventEditComponent = new EventEditComponent(event);
-
-  eventEditComponent.setSaveClickHandler(onReplaceToEvent);
-
-  render(container, eventComponent, Place.BEFOREEND);
-};
-
-const renderDay = (tripDays, day) => {
-  const eventListComponent = new EventsListComponent();
-
-  render(tripDays, new DayComponent(day), Place.BEFOREEND);
-
-  const eventDay = tripDays.querySelector(`#day${day.dayCounter}`);
-
-  render(eventDay, eventListComponent, Place.BEFOREEND);
-
-  const eventList = eventDay.querySelector(`.trip-events__list`);
-
-  eventsCopy.splice(0, getRandomIntegerNumber(5, 1))
-    .forEach((event) => {
-      renderEvent(eventList, event);
-    });
-};
+import {render} from "./utils/render";
 
 const init = () => {
   render(headerInfo, headerInfoComponent, Place.AFTERBEGIN);
@@ -85,11 +19,7 @@ const init = () => {
   render(tripBoard, sortComponent, Place.BEFOREEND);
   render(tripBoard, boardComponent, Place.BEFOREEND);
 
-  const tripDays = document.querySelector(`.trip-days`);
-
-  days.forEach((day) => {
-    renderDay(tripDays, day);
-  });
+  tripController.render(days);
 };
 
 const noEventInit = () => {
@@ -104,8 +34,6 @@ const tripControls = document.querySelector(`.trip-controls`);
 const tripMenuTitle = tripControls.querySelector(`h2`);
 const tripBoard = document.querySelector(`.trip-events`);
 
-const events = generateEvents(EvenOption.COUNT, generateEvent);
-const eventsCopy = events.slice();
 const days = generateEvents(EvenOption.DAY_COUNT, generateDay);
 
 const headerInfoComponent = new HeaderInfoComponent();
@@ -113,8 +41,7 @@ const menuComponent = new MenuComponent();
 const filterComponent = new FilterComponent();
 const sortComponent = new SortComponent();
 const boardComponent = new BoardComponent();
-
-let openEvent;
+const tripController = new TripController(boardComponent);
 
 if (days.length === 0) {
   noEventInit();
